# service/internal/swagger/swagger.yaml
definitions:
  errors.Error:
    properties:
      code:
        type: integer
      detail:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
  schema.Role:
    properties:
      code:
        description: Code of role (unique)
        type: string
      created_at:
        description: Create time
        type: string
      description:
        description: Details about role
        type: string
      id:
        description: Unique ID
        type: string
      menus:
        description: Role menu list
        items:
          $ref: '#/definitions/schema.RoleMenu'
        type: array
      name:
        description: Display name of role
        type: string
      sequence:
        description: Sequence for sorting
        type: integer
      status:
        description: Status of role (disabled, enabled)
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
  schema.RoleForm:
    properties:
      code:
        description: Code of role (unique)
        maxLength: 32
        type: string
      description:
        description: Details about role
        type: string
      menus:
        description: Role menu list
        items:
          $ref: '#/definitions/schema.RoleMenu'
        type: array
      name:
        description: Display name of role
        maxLength: 128
        type: string
      sequence:
        description: Sequence for sorting
        type: integer
      status:
        description: Status of role (enabled, disabled)
        enum:
        - disabled
        - enabled
        type: string
    required:
    - code
    - name
    - status
    type: object
  schema.RoleMenu:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      menu_id:
        description: From Menu.ID
        type: string
      role_id:
        description: From Role.ID
        type: string
      updated_at:
        description: Update time
        type: string
    type: object
...
  schema.User:
    properties:
      created_at:
        description: Create time
        type: string
      email:
        description: Email of user
        type: string
      id:
        description: Unique ID
        type: string
      name:
        description: Name of user
        type: string
      phone:
        description: Phone number of user
        type: string
      remark:
        description: Remark of user
        type: string
      roles:
        description: Roles of user
        items:
          $ref: '#/definitions/schema.UserRole'
        type: array
      status:
        description: Status of user (activated, freezed)
        type: string
      updated_at:
        description: Update time
        type: string
      username:
        description: Username for login
        type: string
    type: object
  schema.UserForm:
    properties:
      email:
        description: Email of user
        maxLength: 128
        type: string
      name:
        description: Name of user
        maxLength: 64
        type: string
      password:
        description: Password for login (md5 hash)
        maxLength: 64
        type: string
      phone:
        description: Phone number of user
        maxLength: 32
        type: string
      remark:
        description: Remark of user
        maxLength: 1024
        type: string
      roles:
        description: Roles of user
        items:
          $ref: '#/definitions/schema.UserRole'
        type: array
      status:
        description: Status of user (activated, freezed)
        enum:
        - activated
        - freezed
        type: string
      username:
        description: Username for login
        maxLength: 64
        type: string
    required:
    - name
    - roles
    - status
    - username
    type: object
  schema.UserRole:
    properties:
      created_at:
        description: Create time
        type: string
      id:
        description: Unique ID
        type: string
      role_id:
        description: From Role.ID
        type: string
      role_name:
        description: From Role.Name
        type: string
      updated_at:
        description: Update time
        type: string
      user_id:
        description: From User.ID
        type: string
    type: object
  util.ResponseResult:
    properties:
      data: {}
      error:
        $ref: '#/definitions/errors.Error'
      success:
        type: boolean
      total:
        type: integer
    type: object
info:
  contact: {}
  description: All packages in llama.cpp whisper.cpp rwkv.cpp chatglm.cpp stable-diffusion.cpp
    provide HTTP APIs through Golang.
  title: godo.cpp
  version: v1.0.0
paths:
...
  /api/v1/roles:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        required: true
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        required: true
        type: integer
      - description: Display name of role
        in: query
        name: name
        type: string
      - description: Status of role (disabled, enabled)
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.Role'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query role list
      tags:
      - RoleAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.RoleForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create role record
      tags:
      - RoleAPI
  /api/v1/roles/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete role record by ID
      tags:
      - RoleAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.Role'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get role record by ID
      tags:
      - RoleAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.RoleForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update role record by ID
      tags:
      - RoleAPI
  /api/v1/users:
    get:
      parameters:
      - default: 1
        description: pagination index
        in: query
        name: current
        required: true
        type: integer
      - default: 10
        description: pagination size
        in: query
        name: pageSize
        required: true
        type: integer
      - description: Username for login
        in: query
        name: username
        type: string
      - description: Name of user
        in: query
        name: name
        type: string
      - description: Status of user (activated, freezed)
        in: query
        name: status
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.User'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Query user list
      tags:
      - UserAPI
    post:
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UserForm'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Create user record
      tags:
      - UserAPI
  /api/v1/users/{id}:
    delete:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Delete user record by ID
      tags:
      - UserAPI
    get:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/util.ResponseResult'
            - properties:
                data:
                  $ref: '#/definitions/schema.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Get user record by ID
      tags:
      - UserAPI
    put:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.UserForm'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Update user record by ID
      tags:
      - UserAPI
  /api/v1/users/{id}/reset-pwd:
    patch:
      parameters:
      - description: unique id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.ResponseResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ResponseResult'
      security:
      - ApiKeyAuth: []
      summary: Reset user password by ID
      tags:
      - UserAPI
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"